deduplicate scheduler dummy classes into shared module
